requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-security-too>=5.5.2
flask-restful==0.3.10
flask-cors==4.0.1
werkzeug==3.0.3
bcrypt==4.1.3
